+++ Changelog.md
@@ -11,7 +11,7 @@
 1. Upgrade the core dependency `react-json-grid` from `v0.7.0` to `v0.9.2`, where several bugs are fixed.
 2. Update the tests to match the new `react-json-grid@0.9.2`.
 3. Make the data routing support the new indexing rule like: `[..., ..., [2]]`, where the last value is an `int` in a one-element list.
-4. Allow the `DashJsonGrid.update_data_by_route` accept a mapping value when updating a column. If all keys in this mapping are `int`, will treat the mapping as the index set.
+4. Allow the `DashJsonGrid.update_data_by_route(...)` to accept a mapping value when updating a column. If all keys in this mapping are `int`, will treat the mapping as the index set.
 
 #### :wrench: Fix
 
@@ -21,6 +21,7 @@
 
 1. Configure `.gitattributes` for specifying the line-breaks of files.
 2. Make the returned column data of `DashJsonGrid.get_data_by_route(...)` or `pop_item_of_object(...)` become `OrderedDict()` if the column data is incomplete.
+3. Improve the compatibility of `dash_json_grid.DashJsonGrid.update_data_by_route(...)`. Now it can update table data even if it contains invalid rows.
 
 ### 0.3.4 @ 10/20/2024